+++ CHANGELOG.md
@@ -4,6 +4,8 @@ UNRELEASED
 
 - Update vendor/gitlab_git to b58c4f436abaf646703bdd80f266fa4c0bab2dd2
   https://gitlab.com/gitlab-org/gitaly/merge_requests/324
+- Add missing cmd.Close in log.GetCommit
+  https://gitlab.com/gitlab-org/gitaly/merge_requests/326
 
 v0.37.0